+++ lib/bitmovin_player.dart
@@ -60,6 +60,7 @@ export 'src/api/player/player_config.dart';
 export 'src/api/player/seek_mode.dart';
 export 'src/api/source/source_config.dart';
 export 'src/api/source/source_options.dart';
+export 'src/api/source/source_remote_control_config.dart';
 export 'src/api/source/source_type.dart';
 export 'src/api/source/timeline_reference_point.dart';
 export 'src/api/ui/fullscreen_handler.dart';